+++ wemake_python_styleguide/violations/refactoring.py
@@ -45,6 +45,7 @@ Summary
    ImplicitDictGetViolation
    ImplicitNegativeIndexViolation
    SimplifiableReturningIfViolation
+   BareRaiseViolation
 
 Refactoring opportunities
 -------------------------
@@ -81,7 +82,7 @@ Refactoring opportunities
 .. autoclass:: ImplicitDictGetViolation
 .. autoclass:: ImplicitNegativeIndexViolation
 .. autoclass:: SimplifiableReturningIfViolation
-
+.. autoclass:: BareRaiseViolation
 """
 
 from typing_extensions import final
@@ -1236,3 +1237,37 @@ class SimplifiableReturningIfViolation(ASTViolation):
 
     error_template = 'Found simplifiable returning `if` condition in a function'
     code = 531
+
+
+@final
+class BareRaiseViolation(ASTViolation):
+    """
+    Forbid bare ``raise`` outside of ``except`` block.
+
+    Reasoning:
+        One could call a function from an ``except`` and have
+        a bare ``raise`` inside but it is considered bad practice
+        to have a bare ``raise`` outside of an ``except`` block.
+
+    Solution:
+        Use a ``raise`` statement inside a parent ``except`` block.
+
+    Example::
+
+        # good
+        def smth():
+            try:
+                ...
+            except:
+                raise
+
+        # bad
+        def smth():
+            raise
+
+    .. versionadded:: 0.15.3
+
+    """
+
+    error_template = 'Bare raise outside of except block detected'
+    code = 532