+++ wemake_python_styleguide/visitors/ast/keywords.py
@@ -32,6 +32,7 @@ from wemake_python_styleguide.violations.consistency import (
     IncorrectYieldFromTargetViolation,
     MultipleContextManagerAssignmentsViolation,
 )
+from wemake_python_styleguide.violations.refactoring import BareRaiseViolation
 from wemake_python_styleguide.visitors.base import BaseNodeVisitor
 from wemake_python_styleguide.visitors.decorators import alias
 
@@ -57,9 +58,11 @@ class WrongRaiseVisitor(BaseNodeVisitor):
 
         Raises:
             RaiseNotImplementedViolation
+            BareRaiseViolation
 
         """
         self._check_exception_type(node)
+        self._check_bare_raise(node)
         self.generic_visit(node)
 
     def _check_exception_type(self, node: ast.Raise) -> None:
@@ -70,6 +73,11 @@ class WrongRaiseVisitor(BaseNodeVisitor):
             self.add_violation(
                 BaseExceptionRaiseViolation(node, text=exception_name),
             )
+
+    def _check_bare_raise(self, node: ast.Raise) -> None:
+        if node.exc is None:
+            if not walk.get_closest_parent(node, ast.ExceptHandler):
+                self.add_violation(BareRaiseViolation(node))
 
 
 @final